requirements.txt
# This requirements file lists all third-party dependencies for this project.
#
# Run 'pip install -r requirements.txt -t lib/' to install these dependencies
# in `lib/` subdirectory.
#
# Note: The `lib` directory is added to `sys.path` by `appengine_config.py`.
#
# Licenses for libraries:
# http://flask.pocoo.org/docs/0.10/license/
# https://github.com/mitsuhiko/werkzeug/blob/master/LICENSE
# https://github.com/lepture/flask-oauthlib/blob/master/LICENSE
# https://github.com/miguelgrinberg/Flask-HTTPAuth/blob/master/LICENSE
# http://docs.python-requests.org/en/latest/user/intro/#apache2Flask==0.10
werkzeug==0.9.4
Flask-OAuthlib>=0.5.0
Flask-HTTPAuth
requests
mock
